% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{ancestors}
\alias{ancestors}
\title{Get the vector of ids of ancestors of a taxon}
\usage{
ancestors(id, taxo, n = Inf)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}

\item{n}{number of levels to look up; n=1 gives the parent, n=2 gives the grand-parent, etc.}
}
\description{
Get the vector of ids of ancestors of a taxon
}
\examples{
db <- src_ecotaxa()
taxo <- extract_taxo(db, ids=c(100,200))
ancestors(200, taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{children}},
  \code{\link{extract_taxo}}, \code{\link{is.taxo}},
  \code{\link{is_leaf}}, \code{\link{lineage}},
  \code{\link{parent}}, \code{\link{taxo_name}}
}
