% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_handle_response.R
\name{api_handle_response}
\alias{api_handle_response}
\title{Handle an API response}
\usage{
api_handle_response(x)
}
\arguments{
\item{x}{an \link[httr:response]{httr::response} object}
}
\description{
Fail, with information, if the status code is not 200. Return the parsed response content otherwise
}
