% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_login.R
\name{api_login}
\alias{api_login}
\title{Log in EcoTaxa}
\usage{
api_login(username, password, cache = TRUE)
}
\arguments{
\item{username}{your EcoTaxa username (i.e. email address).}

\item{password}{your EcoTaxa password, as text.}

\item{cache}{whether to store the token in a file, determined by \code{\link[=api_token_path]{api_token_path()}}. This is \code{TRUE} by default and allows other functions to reuse the token.}
}
\value{
The token string, invisibly. It also stores to token string in the file defined by \code{\link[=api_token_path]{api_token_path()}} if \code{cache=TRUE} (the default).
}
\description{
Log in EcoTaxa
}
\examples{
# log in and get a token
token <- api_login(username="ecotaxa.api.user@gmail.com", password="test!")
token
# check where the token is stored
api_token_path()
readLines(api_token_path())
}
