% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_projects.R
\name{api_project}
\alias{api_project}
\title{Get information for a project}
\usage{
api_project(project_id)
}
\arguments{
\item{project_id}{internal, numeric id of the project.}
}
\value{
A list describing the properties of the project.
}
\description{
Get information for a project
}
\examples{
api_project(185)
}
\seealso{
Other projects: 
\code{\link{api_project_create}()},
\code{\link{api_project_delete}()},
\code{\link{api_project_update}()},
\code{\link{api_projects_filter}()}
}
\concept{projects}
