% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_projects.R
\name{api_project_create}
\alias{api_project_create}
\title{Create a project}
\usage{
api_project_create(title, visible = TRUE, clone_of_id = 0)
}
\arguments{
\item{title}{string, the project title.}

\item{visible}{boolean, when TRUE, the project is visible by all users.}

\item{clone_of_id}{internal, numeric id of a project to clone as a new one. By default it does not clone anything.}
}
\value{
The numeric id of the newly created project.
}
\description{
Create a project and become manager of it. This is reserved to users with the permission to create projects; this is not the case of the default API user, for security reasons.
}
\examples{
\dontrun{
# log in as a user with the permission to create a project
api_login("your@email", "your_password")
# create a project
project_id <- api_project_create("Test project through the API", visible=FALSE)
# get its URL
paste0(sub("api", "prj", api_url()), project_id)
# list it through the API
api_projects_filter("through the API")
# then delete it
api_project_delete(project_id)
api_projects_filter("through the API")
}
}
\seealso{
Other projects: 
\code{\link{api_project_delete}()},
\code{\link{api_projects_filter}()}
}
\concept{projects}
