% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_projects.R
\name{api_project_delete}
\alias{api_project_delete}
\title{Delete a project}
\usage{
api_project_delete(project_id, only_objects = FALSE)
}
\arguments{
\item{project_id}{internal, numeric id of the project.}

\item{only_objects}{when TRUE, delete only the objects and keep the project structure (permissions, free column names, etc.) intact.}
}
\value{
??
}
\description{
Delete a project (of which you are a manager).
}
\examples{
\dontrun{
# log in as a user who is manager on the project to delete
api_login("your@email", "your_password")
# create a project and delete it
project_id <- api_project_create("Test project through the API", visible=FALSE)
api_projects_filter("through the API")
api_project_delete(project_id)
api_projects_filter("through the API")
}
}
\seealso{
Other projects: 
\code{\link{api_project_create}()},
\code{\link{api_project_update}()},
\code{\link{api_projects_filter}()},
\code{\link{api_project}()}
}
\concept{projects}
