% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_projects.R
\name{api_project_update}
\alias{api_project_update}
\title{Update information for a project}
\usage{
api_project_update(project_id, info)
}
\arguments{
\item{project_id}{internal, numeric id of the project.}

\item{info}{a list describing the properties of the project, formatted exactly as the output of api_project().}
}
\value{
NULL upon success.
}
\description{
Update information for a project
}
\examples{
i <- api_project(185)
i$comments
# change comment and update
i$comments <- "This is a test comment"
api_project_update(185, i)
# check that the comment is changed
api_project(185)$comments
# change it back to something else
i$comments <- "No comment"
api_project_update(185, i)
}
\seealso{
Other projects: 
\code{\link{api_project_create}()},
\code{\link{api_project_delete}()},
\code{\link{api_projects_filter}()},
\code{\link{api_project}()}
}
\concept{projects}
