% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_sample.R
\name{api_sample}
\alias{api_sample}
\title{Get information about a single sample}
\usage{
api_sample(id)
}
\arguments{
\item{id}{internal id of the sample}
}
\value{
A nested list with the properties of the sample.
}
\description{
Get information about a single sample
}
\examples{
info <- api_sample(id=15709)
str(info)
}
