% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_token.R
\name{api_token}
\alias{api_token}
\title{Read your EcoTaxa token}
\usage{
api_token(path = api_token_path())
}
\arguments{
\item{path}{path to the file where the token is stored, defined by \verb{[api_token_path()]}.}
}
\value{
The token string.
}
\description{
Read your EcoTaxa token
}
\examples{
# get a token
api_login(username="ecotaxa.api.user@gmail.com", password="test!")
# and now read it
api_token()
}
