% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_token.R
\name{api_token_path}
\alias{api_token_path}
\title{Get the path to your cached EcoTaxa token}
\usage{
api_token_path()
}
\value{
The path to the token file.
}
\description{
This path can be changed with option "ecotaxar.token_path"; by default, it is
"~/.R/ecotaxar/token".
}
\examples{
# get current token path
api_token_path()
# set the option to change the path
options(ecotaxa.token_path="~/.my_ecotaxa_token")
api_token_path()
# NB: write this option in .Rprofile to make it permanent.
}
