% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_update.R
\name{api_update_object_set}
\alias{api_update_object_set}
\title{Batch update metadata of objects}
\usage{
api_update_object_set(ids, updates)
}
\arguments{
\item{ids}{internal ids of the objects to update}

\item{updates}{named list with the metadata field(s) to update and the updated value(s). Due to a current bug, the list should contain either regular fields only (the ones with a grey background in the object page) or free fields only (the ones uploaded by the user, with a regular white background in the object page), but not a mix of both.}
}
\value{
The number of updated items.
}
\description{
Batch update metadata of objects
}
\examples{
# change one field of metadata in one object
api_update_object_set(24473014, list(longitude=10))
# check on this object's page
#   https://ecotaxa.obs-vlfr.fr/objectdetails/24473014

# change several fields in one object
api_update_object_set(24473014, list(longitude=0, latitude=0))
api_update_object_set(24473014, list(area=1, mean=1))

# change several fields in several objects
api_update_object_set(c(24473014, 24473015), list(longitude=5, latitude=-2))
# check on these objects' pages
#   https://ecotaxa.obs-vlfr.fr/objectdetails/24473014
#   https://ecotaxa.obs-vlfr.fr/objectdetails/24473015

# change the taxonomic identification
api_update_object_set(24473014, list(classif_id=1, classif_qual="V"))

# restore
api_update_object_set(
  c(24473014, 24473015),
  list(longitude=7.31567, latitude=43.68500)
) # NB: these objects were actually collected in the same location
api_update_object_set(24473014, list(classif_id=11509, classif_qual="P"))
api_update_object_set(24473014, list(area=1028, mean=224))
api_update_object_set(24473015, list(area=1118, mean=225))
}
