% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url.R
\name{api_url}
\alias{api_url}
\title{Get the base URL of EcoTaxa's API}
\usage{
api_url()
}
\description{
Get the base URL of EcoTaxa's API
}
\details{
The url is "https://ecotaxa.obs-vlfr.fr/api/" by default but can be changed with the option \code{ecotaxa.url}
}
\examples{
api_url()
# set a new default
options(ecotaxa.url="https://ecotaxa.somewhere.edu/api/")
api_url()
# remove the option
options(ecotaxa.url=NULL)
api_url()
}
