% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_users.R
\name{api_user_preference}
\alias{api_user_preference}
\alias{api_get_user_preference}
\alias{api_set_user_preference}
\alias{api_del_user_preference}
\title{Deal with preferences}
\usage{
api_get_user_preference(project_id, key)

api_set_user_preference(project_id, key, value)

api_del_user_preference(project_id, key)
}
\arguments{
\item{project_id}{internal, numeric id of the project.}

\item{key}{string, the name (key) of the preference. Possible preferences are:
\itemize{
\item \code{cwd}: the "current working directory", i.e. last directory on the server from which you imported data in the target project.
\item \code{mru}: the list of taxonomic categories "most recently used" in this project. This list is used in the autocompletion of categories in the classification page.
}}

\item{value}{the value of the preference, when setting it.}
}
\description{
Get, set, or erase a preference for the currently authenticated user (i.e. you) in a project.
}
\examples{
# get preference
api_get_user_preference(185, "cwd")

# set preference and verify it
api_get_user_preference(185, key="foo")
api_set_user_preference(185, key="foo", value="test")
api_get_user_preference(185, key="foo")

# delete preference
api_del_user_preference(185, key="foo")
api_get_user_preference(185, key="foo")

}
\seealso{
Other users: 
\code{\link{api_users_me}()},
\code{\link{api_users_search}()},
\code{\link{api_user}()}
}
\concept{users}
