% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_users.R
\name{api_users}
\alias{api_users}
\title{List EcoTaxa users}
\usage{
api_users()
}
\value{
A list of users, empty when the current logged in user is not application administrator.
}
\description{
This function is accessible to application administrators only.
}
\examples{
api_users()

}
