% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_users.R
\name{api_users_me}
\alias{api_users_me}
\title{Get information about the currently authenticated user (i.e. you)}
\usage{
api_users_me()
}
\value{
A list with user properties, including id, email and name.
}
\description{
Get information about the currently authenticated user (i.e. you)
}
\examples{
api_users_me()

}
