% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{as.list.taxo}
\alias{as.list.taxo}
\title{Convert the taxonomy data.frame into a nested list, for shinyTree}
\usage{
\method{as.list}{taxo}(x, ...)
}
\arguments{
\item{x}{taxonomy data.frame, of class "taxo" (from \code{\link{get_taxo}})}

\item{...}{passed to other methods}
}
\description{
Convert the taxonomy data.frame into a nested list, for shinyTree
}
