% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{as.Node.taxo}
\alias{as.Node.taxo}
\title{Convert a taxonomy data.frame into a Node object, from package data.tree}
\usage{
as.Node.taxo(x)
}
\arguments{
\item{x}{taxonomy data.frame, of class "taxo" (from \code{\link{get_taxo}})}
}
\description{
Convert a taxonomy data.frame into a Node object, from package data.tree
}
