% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{as.taxo}
\alias{as.taxo}
\alias{is.taxo}
\alias{as.list.taxo}
\alias{as.Node.taxo}
\title{Coerce taxonomy data.frames to various formats}
\usage{
as.taxo(x)

is.taxo(x)

\method{as.list}{taxo}(x, ...)

\method{as.Node}{taxo}(x, ...)
}
\arguments{
\item{x}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}

\item{...}{passed to other methods}
}
\description{
Coerce taxonomy data.frames to various formats
}
\examples{
d <- read.csv(text=
"id,parent_id,name
1,NA,living
2,1,fish
3,1,mollusc
4,2,egg
5,3,egg")
is.taxo(d)
d <- as.taxo(d)
is.taxo(d)
as.list(d)
as.Node(d)

db <- src_ecotaxa()
taxo <- extract_taxo(db, c(8000, 20000))
as.Node(taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{children}}, \code{\link{extract_taxo}},
  \code{\link{is_leaf}}, \code{\link{lineage}},
  \code{\link{parent}}, \code{\link{taxo_id}},
  \code{\link{taxo_name}}
}
