% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{children}
\alias{children}
\title{Get id(s) of the children of a taxon}
\usage{
children(id, taxo)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}
}
\description{
Get id(s) of the children of a taxon
}
\examples{
data(taxo)
children(3, taxo)
children(1:4, taxo)
children(c(1:4, NA), taxo)
children(NA, taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{descendants}},
  \code{\link{extract_taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{parent}},
  \code{\link{taxo_id}}, \code{\link{taxo_name}}
}
\concept{taxonomy-related functions}
