% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect_ecotaxa.R
\name{db_connect}
\alias{db_connect}
\alias{db_connect_ecotaxa}
\alias{db_connect_ecopart}
\alias{db_disconnect}
\alias{db_disconnect_ecotaxa}
\alias{db_disconnect_ecopart}
\alias{src_ecotaxa}
\title{Database connection to EcoTaxa}
\usage{
db_connect(host = NULL, dbname = NULL, user = NULL, password = NULL, ...)

db_connect_ecotaxa(...)

db_connect_ecopart(...)

db_disconnect(x)

db_disconnect_ecotaxa(x)

db_disconnect_ecopart(x)

src_ecotaxa()
}
\arguments{
\item{host}{URL of the machine hosting EcoTaxa; when NULL, the default, the function will try the URLs of the various hosts in domain obs-vlfr.fr.}

\item{dbname}{name of the database to connect to.}

\item{user}{name of a user that has read-only access to the database.}

\item{password}{password of that user.}

\item{...}{passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}} for a PostgreSQL connection.}

\item{x}{a database connection created by \code{\link[=db_connect_ecotaxa]{db_connect_ecotaxa()}}.}
}
\value{
An object of class \code{\link[RPostgreSQL:PostgreSQLConnection-class]{RPostgreSQL::PostgreSQLConnection-class()}}.
}
\description{
Connect to and disconnect from the EcoTaxa database (in read-only mode and only from the local network of the machine running EcoTaxa).
}
\examples{
db <- db_connect_ecotaxa()
db
db_disconnect_ecotaxa(db)
# NB: always disconnect after use. Leaving open connections clobbers the server
}
