% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connect_ecotaxa.R
\name{db_connect_ecotaxa}
\alias{db_connect_ecotaxa}
\alias{db_disconnect_ecotaxa}
\alias{src_ecotaxa}
\alias{db_disconnect}
\title{Database connection to EcoTaxa}
\usage{
db_connect_ecotaxa()

db_disconnect_ecotaxa(x)

src_ecotaxa()

db_disconnect(x)
}
\arguments{
\item{x}{a database connection created by \code{\link[=db_connect_ecotaxa]{db_connect_ecotaxa()}}}
}
\value{
An object of class
}
\description{
Connect to and disconnect from the EcoTaxa database (in read-only mode).
}
\examples{
db <- db_connect_ecotaxa()
db
db_disconnect_ecotaxa(db)
# NB: always disconnect after use. Leaving open connections clobbers the server
}
