% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{db_objects}
\alias{db_objects}
\title{Get objects and associated metadata}
\usage{
db_objects(
  db,
  projids,
  ...,
  object_fields = NULL,
  process_fields = NULL,
  acquis_fields = NULL,
  sample_fields = NULL
)
}
\arguments{
\item{db}{connection to the database, from \verb{[db_connect_ecotaxa()]}.}

\item{projids}{numeric identifiers of projects to extract data from.}

\item{...}{additional filters on the objects table. Typical filters are on:
\itemize{
\item \code{classif_qual}: "V" for validated, "P" for predicted, "D" for dubious,
\item \code{classif_id}: numeric ids of the taxa of interest,
\item \code{sampleid}: internal numeric identifiers of the samples (from the "samples" table).
}}

\item{*_fields}{vectors of column names to extract in the objects, sample, acquisitions, and process tables. NULL (the default) extracts nothing, "all" extracts all fields.}
}
\value{
A tibble with columns that are systematically extracted and the additional metadata requested
}
\description{
Get objects and associated metadata
}
\examples{
db <- db_connect_ecotaxa()
# extract validated objects from a couple projects with some metadata
d <- extract_objects(
  db, projids=c(658, 756), classif_qual=="V",
  object_fields=c("area", "major"),
  process_fields="particle_pixel_size_mm",
  acquis_fields="sub_part",
  sample_fields="tot_vol"
)
# extract all objects from a project with all object-level metadata
d <- extract_objects(
  db, projids=658,
  object_fields="all"
)
}
