% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxo.R
\name{extract_taxo}
\alias{extract_taxo}
\title{Extract a subset of the taxonomy for a set of ids}
\usage{
extract_taxo(db, ids, recursive = TRUE)
}
\arguments{
\item{db}{a connection to the EcoTaxa database (from \code{\link{src_ecotaxa}})}

\item{ids}{vector of numerical ids of taxonomic classes}

\item{recursive}{when \code{TRUE}, extract the full taxonomy, until the root}
}
\description{
Extract a subset of the taxonomy for a set of ids
}
\examples{
db <- src_ecotaxa()
extract_taxo(db, 84000:84003)
extract_taxo(db, 84000:84003, recursive=FALSE)
}
