% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_tree_operations.R
\name{is_leaf}
\alias{is_leaf}
\title{Tests wether an id corresponds to a leaf in a taxonomic tree}
\usage{
is_leaf(id, taxo)
}
\description{
Tests wether an id corresponds to a leaf in a taxonomic tree
}
