% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_names}
\alias{map_names}
\title{Parse a mapping specification and rename columns}
\usage{
map_names(x, mapping)
}
\arguments{
\item{x}{a data.frame with un-mapped columns (in the form "n01", "n02", etc.)}

\item{mapping}{an EcoTaxa mapping string}
}
\description{
Parse a mapping specification and rename columns
}
\examples{
d <- data.frame(id=c(1234, 1235), n01=c(12,34), n02=c(125,156))
mapping <- "n01=area\\nn02=mean\\nn03=stddev"
map_names(d, mapping)
}
