% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{parent}
\alias{parent}
\alias{parents}
\title{Get id of the parent of some taxa}
\usage{
parent(taxon_ids, taxo)

parents(taxon_ids, taxo)
}
\arguments{
\item{taxon_ids}{internal, numeric ids of the taxa.}

\item{taxo}{a taxonomy data.frame, typically from \code{\link[=extract_taxo]{extract_taxo()}}.}
}
\value{
A vector of taxon ids, as long as the input, containing the id of the parent of each element in the input.
}
\description{
Get id of the parent of some taxa
}
\details{
\code{parent} and \code{parents} are just synonyms. The function works per taxon but is vectorised over \code{taxon_ids}.
}
\examples{
print(as.Node(taxo), "id")
parent(3, taxo)
parent(1:4, taxo)
parent(c(5, NA), taxo)
\dontrun{ parent(12, taxo) }
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{as.taxo}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{taxo_id}()},
\code{\link{taxo_name}()}
}
\concept{taxonomy-related functions}
