% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{parse_mapping}
\alias{parse_mapping}
\title{Parse a mapping specification into a named vector}
\usage{
parse_mapping(mapping)
}
\arguments{
\item{mapping}{an EcoTaxa mapping string}
}
\description{
Parse a mapping specification into a named vector
}
\examples{
mapping <- "n01=area\\nn02=mean\\nn03=stddev"
parse_mapping(mapping)
}
