% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url.R
\name{query_string}
\alias{query_string}
\title{Build a query with parameters}
\usage{
query_string(...)
}
\arguments{
\item{...}{named arguments to be turned into query parameters.}
}
\description{
Build a query with parameters
}
\examples{
query_string(foo="a", bar=NULL, bob=1)
}
