% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecotaxa_tsv.R
\name{read_ecotaxa_tsv}
\alias{read_ecotaxa_tsv}
\alias{read_etx}
\title{Read tsv file from EcoTaxa}
\usage{
read_ecotaxa_tsv(file, col_types = "auto", ...)

read_etx(file, col_types = "auto", ...)
}
\arguments{
\item{file}{path to the tsv file}

\item{col_types}{a column specification for \code{\link[readr]{read_tsv}}, build with \code{\link[readr]{cols}}, or "auto" for automatic parsing}

\item{...}{passed to \code{\link[readr]{read_tsv}}}
}
\description{
Read tsv file from EcoTaxa
}
\examples{
d <- read_ecotaxa_tsv(file=system.file("extdata", "export.tsv", package="ecotaxar"))
glimpse(d)
d <- read_ecotaxa_tsv(file=system.file("extdata", "import.tsv", package="ecotaxar"))
glimpse(d)
}
