% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.Node}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.tree}{\code{\link[data.tree]{as.Node}}}
}}

