% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove empty columns from a data.frame}
\usage{
remove_empty_cols(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Useful to remove unmapped columns extracted from EcoTaxa, which are all empty
}
\examples{
x <- data.frame(a=c(1,NA), b=NA, c=c(NA, "foo"))
remove_empty_cols(x)
}
