% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_ecotaxa.R
\name{src_ecotaxa}
\alias{src_ecotaxa}
\alias{db_disconnect}
\title{Database connection to EcoTaxa}
\usage{
src_ecotaxa()

db_disconnect(x)
}
\description{
Connect to and disconnect from the EcoTaxa database (in read-only mode).
}
\examples{
db <- src_ecotaxa()
db
db_disconnect(db)
# NB: always disconnect after use. Leaving open connections clobbers the server
}
