% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\docType{data}
\name{taxo}
\alias{taxo}
\title{Minimal, fake taxonomy table}
\format{
A data frame with 7 rows and 3 variables:
\describe{
\item{id}{taxon identifier}
\item{parent_id}{identifier of the parent}
\item{name}{name of the taxon (i.e. text)}
}
}
\usage{
taxo
}
\description{
Minimal, fake taxonomy table
}
\keyword{datasets}
