% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_conversions.R
\name{taxo_name}
\alias{taxo_name}
\alias{lineage}
\alias{taxo_id}
\title{Get the names of taxa from their ids}
\usage{
taxo_name(ids, taxo)

lineage(id, taxo)

taxo_id(names, taxo)
}
\arguments{
\item{ids}{vector of numeric taxonomic ids}

\item{taxo}{taxonomy data.frame, with columns \code{id} and \code{name} at least, typically from \code{\link{get_taxo}}}

\item{id}{numeric taxonomic id}

\item{names}{vector of character strings with taxonomic names}
}
\description{
Get the names of taxa from their ids

Get the full lineage of a taxon from its id

Get the ids of taxa from their names
}
