% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_tree_operations.R
\name{ancestors}
\alias{ancestors}
\title{Get the vector of ids of ancestors of a taxon}
\usage{
ancestors(id, taxo, n = Inf)
}
\arguments{
\item{id}{numeric ids of taxa (typically a single one)}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}

\item{n}{number of levels to look up; n=1 gives the parent, n=2 gives the grand-parent, etc.}
}
\description{
Get the vector of ids of ancestors of a taxon
}
