% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{ancestors}
\alias{ancestors}
\title{Get the ids of all ancestors of some taxa}
\usage{
ancestors(taxon_ids, taxo, n = Inf)
}
\arguments{
\item{taxon_ids}{internal, numeric ids of the taxa.}

\item{taxo}{a taxonomy data.frame, typically from \code{\link[=extract_taxo]{extract_taxo()}}.}

\item{n}{maximum number of levels to look up; n=1 gives the parents, n=2 gives the parents and the grand-parents, etc.}
}
\value{
A vector of taxonomic ids containing the input \code{taxon_ids} and all their ancestors. With a single taxon in \code{taxon_ids}, the ancestors are in ordered from the taxon to the root of the tree: they define the lineage. When \code{taxon_ids} contains multiple taxa, the order is non-trivial.
}
\description{
Get the ids of all ancestors of some taxa
}
\details{
Even with \code{n=1}, the function is different from \code{\link[=parent]{parent()}} because it returns all unique parents for all elements of \code{taxon_ids}, not one parent per element.
}
\examples{
print(as.Node(taxo), "id")
ancestors(5, taxo)
ancestors(1, taxo)
ancestors(NA, taxo)
ancestors(1:7, taxo)
ancestors(6, taxo, n=1)
ancestors(6, taxo, n=2)
ancestors(6, taxo, n=3)
ancestors(6, taxo, n=10)
# NB:
ancestors(5:6, taxo, n=1)
parent(5:6, taxo)
}
\seealso{
Other taxonomy-related functions: 
\code{\link{as.taxo}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{parent}()},
\code{\link{taxo_id}()},
\code{\link{taxo_name}()}
}
\concept{taxonomy-related functions}
