% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{ancestors}
\alias{ancestors}
\title{Get ids of ancestors of a taxon}
\usage{
ancestors(id, taxo, n = Inf)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}

\item{n}{number of levels to look up; n=1 gives the parents, n=2 gives the grand-parents, etc.}
}
\description{
Get ids of ancestors of a taxon
}
\details{
Even with \code{n=1} function is different from \code{[parent()]} because it returns a vector of unique parents for all input \code{id} values, not necessarily in order.
}
\examples{
data(taxo)
ancestors(6, taxo)
ancestors(5:7, taxo)
ancestors(1, taxo)
ancestors(NA, taxo)
ancestors(6, taxo, n=1)
ancestors(6, taxo, n=2)
ancestors(6, taxo, n=3)
ancestors(6, taxo, n=10)
# NB:
ancestors(5:7, taxo, n=1)
parent(5:7, taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{as.taxo}},
  \code{\link{children}}, \code{\link{descendants}},
  \code{\link{extract_taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{parent}},
  \code{\link{taxo_id}}, \code{\link{taxo_name}}
}
