% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_object.R
\name{api_object}
\alias{api_object}
\title{Get information about a single object}
\usage{
api_object(object_id)
}
\arguments{
\item{object_id}{internal, numeric id of the object.}
}
\value{
A nested list with the properties of the object.
}
\description{
Get information about a single object
}
\examples{
info <- api_object(24473014)
str(info, 1)
}
