% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_taxonomy.R
\name{api_taxon}
\alias{api_taxon}
\title{Get information about a single taxon}
\usage{
api_taxon(taxon_id)
}
\arguments{
\item{taxon_id}{internal, numeric id of the taxon.}
}
\value{
A list containing
\itemize{
\item \code{id} : the numerical id of the taxon,
\item \code{display_name} : a unique name for the taxon (in which the parent name gets appended if the taxon name is not unique),
\item \code{lineage} : a vector of the taxon name and its ancestors, in order of ascendance.
}
}
\description{
Get information about a single taxon
}
\examples{
api_taxon(1234)
api_taxon(85004)
api_taxon(84968)
}
