% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{as.taxo}
\alias{as.taxo}
\alias{is.taxo}
\alias{as.list.taxo}
\alias{as.Node.taxo}
\title{Coerce a taxonomy data.frame to various formats}
\usage{
as.taxo(x)

is.taxo(x)

\method{as.list}{taxo}(x, ...)

\method{as.Node}{taxo}(x, ...)
}
\arguments{
\item{x}{a taxonomy data.frame, typically from \code{\link[=extract_taxo]{extract_taxo()}}.}

\item{...}{passed to other methods.}
}
\description{
Coerce a taxonomy data.frame to various formats
}
\details{
A taxonomy data.frame is just a regular data.frame with columns \code{id}, \code{parent_id}, and \code{name}. The links between the \code{id} and \code{parent_id} define the taxonomic hierarchy.
It can be converted into an actual tree using \code{\link[data.tree:as.Node]{data.tree::as.Node()}} and into a nested list.
}
\examples{
taxo
as.Node(taxo)
print(as.Node(taxo), "id", "parent_id")
as.list(taxo)

# transforming into a taxonomy data.frame is just a mater of adding a class
df <- taxo
class(df)
is.taxo(df)
df <- as.data.frame(df)
class(df)
is.taxo(df)
df <- as.taxo(df)
class(df)
is.taxo(df)

\dontrun{
db <- db_connect_ecotaxa()
taxo <- extract_taxo(db, c(8000, 20000))
as.Node(taxo)
}
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{parent}()},
\code{\link{taxo_id}()},
\code{\link{taxo_name}()}
}
\concept{taxonomy-related functions}
