% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors_children.R
\name{children}
\alias{children}
\title{Get the vector if ids of children of a taxon}
\usage{
children(id, taxo, n = Inf)
}
\arguments{
\item{id}{numeric ids of taxa (typically a single one)}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}

\item{n}{number of levels to look down; n=1 gives the direct children, n=2 gives grand children (i.e. children of all children), etc.}
}
\description{
Get the vector if ids of children of a taxon
}
