% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{children}
\alias{children}
\title{Get ids of the children of some taxa}
\usage{
children(taxon_ids, taxo, simplify = TRUE)
}
\arguments{
\item{taxon_ids}{internal, numeric ids of the taxa.}

\item{taxo}{a taxonomy data.frame, typically from \code{\link[=extract_taxo]{extract_taxo()}}.}

\item{simplify}{when \code{taxon_ids} contains only one taxon and \code{simplify} is \code{TRUE}, return the vector of ids of the children directly, not a list of length one.}
}
\value{
A list of vectors of taxon ids, as long as the input, containing the ids of the children of each element in \code{taxon_ids}.
}
\description{
Get ids of the children of some taxa
}
\details{
The function works per taxon but is vectorised over \code{taxon_ids}.
}
\examples{
print(as.Node(taxo), "id")
children(3, taxo)
children(1:4, taxo)
children(c(5, NA), taxo)
\dontrun{ children(12, taxo) }
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{as.taxo}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{parent}()},
\code{\link{taxo_id}()},
\code{\link{taxo_name}()}
}
\concept{taxonomy-related functions}
