% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{descendants}
\alias{descendants}
\title{Get ids of descendants of a taxon}
\usage{
descendants(id, taxo, n = Inf)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}

\item{@param}{n number of levels to look down; n=1 gives the direct children, n=2 gives grand-children (i.e. children of all children), etc.}
}
\description{
Get ids of descendants of a taxon
}
\examples{
taxo
descendants(3, taxo)
descendants(1:3, taxo)
descendants(7, taxo)
descendants(NA, taxo)
descendants(1, taxo, n=1)
descendants(1, taxo, n=2)
descendants(1, taxo, n=3)
descendants(1, taxo, n=10)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{children}},
  \code{\link{extract_taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{parent}},
  \code{\link{taxo_id}}, \code{\link{taxo_name}}
}
