% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{extract_taxo}
\alias{extract_taxo}
\title{Extract a subset of the taxonomy for a set of ids}
\usage{
extract_taxo(db, taxon_ids, recursive = TRUE)
}
\arguments{
\item{db}{a connection to an EcoTaxa database (from \code{\link[=db_connect_ecotaxa]{db_connect_ecotaxa()}}).}

\item{taxon_ids}{internal, numeric ids of the taxa.}

\item{recursive}{when \code{TRUE}, extract the full taxonomy, until the root.}
}
\description{
Extract a subset of the taxonomy for a set of ids
}
\examples{
db <- db_connect_ecotaxa()
extract_taxo(db, 84000:84003)
extract_taxo(db, 84000:84003, recursive=FALSE)
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{as.taxo}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{parent}()},
\code{\link{taxo_id}()},
\code{\link{taxo_name}()}
}
\concept{taxonomy-related functions}
