% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxo.R
\name{get_taxo}
\alias{get_taxo}
\title{Extract a subset of the taxonomy for a set of ids}
\usage{
get_taxo(db, ids, recursive = FALSE)
}
\arguments{
\item{db}{a connection to the EcoTaxa database (from \code{\link{src_ecotaxa}})}

\item{ids}{vector of numerical ids of classes}

\item{recursive}{when \code{TRUE}, extract the full taxonomy, until the root}
}
\description{
Extract a subset of the taxonomy for a set of ids
}
