% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{is.taxo}
\alias{is.taxo}
\alias{as.list.taxo}
\alias{as.Node.taxo}
\title{Coerce taxonomy data.frames to various formats}
\usage{
is.taxo(x)

\method{as.list}{taxo}(x, ...)

\method{as.Node}{taxo}(x)
}
\arguments{
\item{x}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}

\item{...}{passed to other methods}
}
\description{
Coerce taxonomy data.frames to various formats
}
\examples{
db <- src_ecotaxa()
taxo <- extract_taxo(db, c(10,20))
is.taxo(taxo)
as.list(taxo)
as.Node(taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{children}}, \code{\link{extract_taxo}},
  \code{\link{is_leaf}}, \code{\link{lineage}},
  \code{\link{parent}}, \code{\link{taxo_name}}
}
