% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{is_leaf}
\alias{is_leaf}
\title{Tests whether taxa are leaves in a taxonomic tree}
\usage{
is_leaf(taxon_ids, taxo)
}
\arguments{
\item{taxon_ids}{internal, numeric ids of the taxa.}

\item{taxo}{a taxonomy data.frame, typically from \code{\link[=extract_taxo]{extract_taxo()}}.}
}
\description{
Tests whether taxa are leaves in a taxonomic tree
}
\examples{
print(as.Node(taxo), "id")
is_leaf(5, taxo)
is_leaf(6, taxo)
is_leaf(NA, taxo)
is_leaf(c(5,NA,6,NA), taxo)
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{as.taxo}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{lineage}()},
\code{\link{parent}()},
\code{\link{taxo_id}()},
\code{\link{taxo_name}()}
}
\concept{taxonomy-related functions}
