% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{is_leaf}
\alias{is_leaf}
\title{Tests wether an id corresponds to a leaf in a taxonomic tree}
\usage{
is_leaf(id, taxo)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}
}
\description{
Tests wether an id corresponds to a leaf in a taxonomic tree
}
\examples{
db <- src_ecotaxa()
taxo <- extract_taxo(db, ids=c(100,200))
taxo
is_leaf(3, taxo)
is_leaf(200, taxo)
is_leaf(c(3, 200), taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{children}}, \code{\link{extract_taxo}},
  \code{\link{is.taxo}}, \code{\link{lineage}},
  \code{\link{parent}}, \code{\link{taxo_id}},
  \code{\link{taxo_name}}
}
