% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{is_leaf}
\alias{is_leaf}
\title{Tests whether a taxon is a leaf in a taxonomic tree}
\usage{
is_leaf(id, taxo)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}
}
\description{
Tests whether a taxon is a leaf in a taxonomic tree
}
\examples{
taxo
is_leaf(3, taxo)
is_leaf(6, taxo)
is_leaf(NA, taxo)
is_leaf(c(5,NA,6,NA), taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{children}},
  \code{\link{descendants}}, \code{\link{extract_taxo}},
  \code{\link{lineage}}, \code{\link{parent}},
  \code{\link{taxo_id}}, \code{\link{taxo_name}}
}
