% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{lineage}
\alias{lineage}
\title{Get the full lineage of a taxon}
\usage{
lineage(id, taxo, rooted = FALSE)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}

\item{rooted}{when \code{TRUE}, add a root (#) to the tree}
}
\description{
Get the full lineage of a taxon
}
\examples{
daya(taxo)
lineage(6, taxo)
lineage(6, taxo, rooted=TRUE)
lineage(c(6, 7), taxo)
lineage(NA, taxo)
lineage(c(6, NA), taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{children}},
  \code{\link{descendants}}, \code{\link{extract_taxo}},
  \code{\link{is_leaf}}, \code{\link{parent}},
  \code{\link{taxo_id}}, \code{\link{taxo_name}}
}
\concept{taxonomy-related functions}
