% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{lineage}
\alias{lineage}
\title{Get the full lineage of a taxon from its id}
\usage{
lineage(ids, taxo, rooted = FALSE)
}
\arguments{
\item{ids}{vector of numerical ids of taxonomic classes}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}

\item{rooted}{when \code{TRUE}, add a root (#) to the tree}
}
\description{
Get the full lineage of a taxon from its id
}
\examples{
db <- src_ecotaxa()
taxo <- extract_taxo(db, ids=c(100,200))
lineage(200, taxo)
lineage(c(3, 200), taxo)
lineage(200, taxo, rooted=T)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{children}}, \code{\link{extract_taxo}},
  \code{\link{is.taxo}}, \code{\link{is_leaf}},
  \code{\link{parent}}, \code{\link{taxo_name}}
}
