% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{parent}
\alias{parent}
\title{Get id of the parent of a taxon}
\usage{
parent(id, taxo)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}
}
\description{
Get id of the parent of a taxon
}
\examples{
db <- src_ecotaxa()
taxo <- extract_taxo(db, ids=c(100,200))
taxo
parent(3, taxo)
parent(c(200, 3), taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{children}},
  \code{\link{extract_taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{taxo_id}},
  \code{\link{taxo_name}}
}
