% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{parent}
\alias{parent}
\alias{parents}
\title{Get id of the parent of a taxon}
\usage{
parent(id, taxo)

parents(id, taxo)
}
\arguments{
\item{id}{numerical ids of taxonomic classes (typically a single one)}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}
}
\description{
Get id of the parent of a taxon
}
\examples{
data(taxo)
parent(3, taxo)
parent(1:4, taxo)
parent(c(5, NA), taxo)
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{as.taxo}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{taxo_id}()},
\code{\link{taxo_name}()}
}
\concept{taxonomy-related functions}
