% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{taxo_id}
\alias{taxo_id}
\title{Get the ids of taxa from their names}
\usage{
taxo_id(names, taxo)
}
\arguments{
\item{names}{vector of character strings with taxonomic names}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}
}
\description{
Get the ids of taxa from their names
}
\examples{
db <- src_ecotaxa()
taxo <- extract_taxo(db, ids=c(100,200))
taxo
taxo_id("Foo", taxo)
taxo_id("Euzebyaceae", taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{children}},
  \code{\link{extract_taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{parent}},
  \code{\link{taxo_name}}
}
