% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{taxo_name}
\alias{taxo_name}
\title{Get the names of taxa from their ids}
\usage{
taxo_name(ids, taxo)
}
\arguments{
\item{ids}{vector of numerical ids of taxonomic classes}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}
}
\description{
Get the names of taxa from their ids
}
\examples{
db <- src_ecotaxa()
taxo <- extract_taxo(db, ids=c(100,200))
taxo
taxo_name(200, taxo)
taxo_name(1, taxo)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{children}},
  \code{\link{extract_taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{parent}},
  \code{\link{taxo_id}}
}
