% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{taxo_name}
\alias{taxo_name}
\alias{taxo_id}
\title{Get the names of taxa from their ids}
\usage{
taxo_name(ids, taxo)

taxo_id(names, taxo)
}
\arguments{
\item{ids}{vector of numerical ids of taxonomic classes}

\item{taxo}{taxonomy data.frame, with columns `id` and `parent_id` at least, typically from from \code{\link{get_taxo}} with \code{recursive=TRUE}}

\item{names}{vector of character strings with taxonomic names}
}
\description{
Get the names of taxa from their ids

Get the ids of taxa from their names
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{children}}, \code{\link{extract_taxo}},
  \code{\link{is.taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{parent}}
}
