% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{taxo_name}
\alias{taxo_name}
\title{Get the names of taxa from their ids}
\usage{
taxo_name(taxon_ids, taxo, unique = FALSE)
}
\arguments{
\item{taxon_ids}{internal, numeric ids of the taxa.}

\item{taxo}{a taxonomy data.frame, typically from \code{\link[=extract_taxo]{extract_taxo()}}.}

\item{unique}{force names to be unique by adding the parent name when needed.}
}
\value{
A vector names, one for each element of \code{taxon_ids}.
}
\description{
Get the names of taxa from their ids
}
\examples{
print(as.Node(taxo), "id")
taxo_name(5, taxo)
taxo_name(2:7, taxo)
taxo_name(2:7, taxo, unique=TRUE)
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{as.taxo}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{parent}()},
\code{\link{taxo_id}()}
}
\concept{taxonomy-related functions}
