% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{taxo_name}
\alias{taxo_name}
\title{Get the names of taxa from their ids}
\usage{
taxo_name(ids, taxo, unique = FALSE, human_readable = TRUE)
}
\arguments{
\item{ids}{vector of numerical ids of taxonomic classes}

\item{taxo}{a taxonomy data.frame, typically from \code{\link{extract_taxo}}}

\item{unique}{force names to be unique by adding the parent name when needed}

\item{human_readable}{when names are made unique and this is TRUE, add the parent name in parentheses after the taxon name (like on EcoTaxa); when FALSE, use `parent name`_`taxon name`.}
}
\description{
Get the names of taxa from their ids
}
\examples{
taxo <- read.csv(text=
"id,parent_id,name
1,NA,living
2,1,fish
3,1,mollusc
4,2,egg
5,3,egg")
taxo <- as.taxo(taxo)
as.Node(taxo)
taxo_name(5, taxo)
taxo_name(1:3, taxo)
taxo_name(3:5, taxo)
taxo_name(3:5, taxo, unique=TRUE)
}
\seealso{
Other taxonomy-related functions: \code{\link{ancestors}},
  \code{\link{as.taxo}}, \code{\link{children}},
  \code{\link{extract_taxo}}, \code{\link{is_leaf}},
  \code{\link{lineage}}, \code{\link{parent}},
  \code{\link{taxo_id}}
}
