% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo.R
\name{taxo_name}
\alias{taxo_name}
\title{Get the names of taxa from their ids}
\usage{
taxo_name(id, taxo, unique = FALSE)
}
\arguments{
\item{taxo}{a taxonomy data.frame, typically from \code{\link[=extract_taxo]{extract_taxo()}}.}

\item{unique}{force names to be unique by adding the parent name when needed}
}
\description{
Get the names of taxa from their ids
}
\examples{
data(taxo)
taxo_name(5, taxo)
taxo_name(2:6, taxo)
taxo_name(2:6, taxo, unique=TRUE)
}
\seealso{
Other taxonomy-related functions: 
\code{\link{ancestors}()},
\code{\link{as.taxo}()},
\code{\link{children}()},
\code{\link{descendants}()},
\code{\link{extract_taxo}()},
\code{\link{is_leaf}()},
\code{\link{lineage}()},
\code{\link{parent}()},
\code{\link{taxo_id}()}
}
\concept{taxonomy-related functions}
