% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ecotaxa.R
\name{tbl_ecotaxa}
\alias{tbl_ecotaxa}
\title{Connect to EcoTaxa tables}
\usage{
tbl_ecotaxa(db)
}
\arguments{
\item{db}{a connection to the EcoTaxa database (from \code{\link{src_ecotaxa}})}
}
\description{
Connect to EcoTaxa tables
}
